+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.0-beta.2
+
+- Update settings tab UI and default settings
+- Add update notifications and funding info
+
 ## 0.2.0-beta.1
 
 - Rewrote all major parts in rust